Fix circular dependency in checkout session imports
Moved createInitialInvoiceForPurchase from bookkeeping.ts to a new
bookkeeping/invoices.ts file to break the circular dependency that was
causing editCheckoutSession to be undefined when imported.

The circular dependency was:
- checkoutSessions.ts imports from bookkeeping.ts
- This caused module initialization issues making exports undefined

Changes:
- Created new bookkeeping/invoices.ts with createInitialInvoiceForPurchase
- Updated checkoutSessions.ts to import from ./invoices instead of ../bookkeeping
- Updated bookkeeping.ts to re-export from new location for backward compatibility
- Removed unused import from clearDiscountCode.ts
